created applicationList to be more modular and easier to understand
