+++ frontend/components/employeeComponents/applicationPagination.tsx
@@ -1,0 +1,48 @@
+import {
+    Pagination,
+    PaginationContent,
+    PaginationItem,
+    PaginationLink,
+    PaginationNext,
+    PaginationPrevious,
+  } from "@/components/ui/pagination"
+
+type PaginationProps = {
+  currentPage: number;
+  totalPages: number;
+  onPageChange: (prev: number) => void;
+};
+
+export const ApplicationPagination = ({currentPage, totalPages, onPageChange}: PaginationProps)=>{
+    return <>
+          <Pagination>
+        <PaginationContent>
+          <PaginationItem>
+            <PaginationPrevious 
+              href="#" 
+              onClick={() => onPageChange(Math.max(currentPage - 1, 1))}
+              className={currentPage === 1 ? 'pointer-events-none opacity-50' : ''}
+            />
+          </PaginationItem>
+          {[...Array(totalPages)].map((_, index) => (
+            <PaginationItem key={index}>
+              <PaginationLink 
+                href="#" 
+                onClick={() => onPageChange(index + 1)}
+                isActive={currentPage === index + 1}
+              >
+                {index + 1}
+              </PaginationLink>
+            </PaginationItem>
+          ))}
+          <PaginationItem>
+            <PaginationNext 
+              href="#" 
+              onClick={() => onPageChange(Math.min(currentPage + 1, totalPages))}
+              className={currentPage === totalPages ? 'pointer-events-none opacity-50' : ''}
+            />
+          </PaginationItem>
+        </PaginationContent>
+      </Pagination>
+    </>
+}
